frontend: rewrite hotkey HOC
